+++ src/django/api/views.py
@@ -80,7 +80,9 @@ from api.models import (FacilityList,
                         Version,
                         FacilityLocation,
                         Source,
-                        ApiBlock)
+                        ApiBlock,
+                        EmbedConfig,
+                        EmbedField)
 from api.processing import (parse_csv_line,
                             parse_csv,
                             parse_excel,
@@ -107,7 +109,8 @@ from api.serializers import (FacilityListSerializer,
                              LogDownloadQueryParamsSerializer,
                              FacilityUpdateLocationParamsSerializer,
                              ApiBlockSerializer,
-                             FacilityActivityReportSerializer)
+                             FacilityActivityReportSerializer,
+                             EmbedConfigSerializer)
 from api.countries import COUNTRY_CHOICES
 from api.aws_batch import submit_jobs
 from api.permissions import IsRegisteredAndConfirmed, IsAllowedHost
@@ -3489,3 +3492,87 @@ class ContributorFacilityListViewSet(viewsets.ReadOnlyModelViewSet):
         ]
 
         return Response(response_data)
+
+
+class EmbedConfigAutoSchema(AutoSchema):
+    def get_link(self, path, method, base_url):
+        return None
+
+
+def create_embed_fields(fields_data, embed_config):
+    if len(fields_data) != len(set([f['order'] for f in fields_data])):
+        raise ValidationError('Fields cannot have the same order.')
+
+    for field_data in fields_data:
+        EmbedField.objects.create(embed_config=embed_config, **field_data)
+
+
+@schema(EmbedConfigAutoSchema())
+class EmbedConfigViewSet(mixins.ListModelMixin,
+                         mixins.RetrieveModelMixin,
+                         mixins.UpdateModelMixin,
+                         mixins.CreateModelMixin,
+                         viewsets.GenericViewSet):
+    """
+    View EmbedConfig.
+    """
+    queryset = EmbedConfig.objects.all()
+    serializer_class = EmbedConfigSerializer
+
+    def get_contributor(self, request):
+        try:
+            contributor_id = request.user.contributor.id
+            contributor = Contributor.objects.get(id=contributor_id)
+        except Contributor.DoesNotExist:
+            raise ValidationError('Contributor not found for requesting user.')
+
+        return contributor
+
+    @transaction.atomic
+    def create(self, request):
+        if not request.user.is_authenticated:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+        contributor = self.get_contributor(request)
+        if contributor.embed_config is not None:
+            raise ValidationError(
+                'Contributor has an existing embed configuration.')
+
+        fields_data = request.data.pop('embed_fields', [])
+
+        serializer = self.get_serializer(data=request.data)
+
+        if serializer.is_valid(raise_exception=True):
+            serializer.save()
+
+            embed_config = EmbedConfig.objects.get(id=serializer.data['id'])
+
+            # Assign embed config to contributor
+            contributor.embed_config = embed_config
+            contributor.save()
+
+            create_embed_fields(fields_data, embed_config)
+
+            response_data = self.get_serializer(embed_config).data
+
+            return Response(response_data)
+
+    @transaction.atomic
+    def update(self, request, pk=None):
+        if not request.user.is_authenticated:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+        contributor = self.get_contributor(request)
+
+        embed_config = EmbedConfig.objects.get(id=pk)
+
+        if embed_config.contributor.id is not contributor.id:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
+        fields_data = request.data.pop('embed_fields', [])
+
+        # Update field data by deleting and recreating
+        EmbedField.objects.filter(embed_config=embed_config).delete()
+        create_embed_fields(fields_data, embed_config)
+
+        return super(EmbedConfigViewSet, self).update(request, pk=pk)